requirements.txt
discord.py~=2.4.0
aiomysql~=0.2.0
python-dotenv~=1.0.1
sqlalchemy~=2.0.31
discord-webhook~=1.3.1
cryptography~=43.0.0
pymongo~=4.8.0